requirements.txt
fastapi
uvicorn

SQLAlchemy
asyncpg

pydantic
pydantic-settings

alembic

pytest
pytest-asyncio
httpx

# Дополнено по заданию

python-jose[cryptography]
passlib[bcrypt]

passlib==1.7.4
bcrypt==4.0.1

asgi-lifespan
pytest-anyio
anyio